rename `program` module to `elgamal-program`
