+++ zk-sdk/src/elgamal_program/proof_data/errors.rs
